+++ hyena-core/src/main/java/io/github/alphajiang/hyena/ds/mapper/PointRecMapper.java
@@ -23,6 +23,7 @@ import io.github.alphajiang.hyena.model.po.PointRecPo;
 import org.apache.ibatis.annotations.Mapper;
 import org.apache.ibatis.annotations.Param;
 
+import java.math.BigDecimal;
 import java.util.Date;
 import java.util.List;
 
@@ -49,10 +50,10 @@ public interface PointRecMapper {
                         @Param(value = "rec") PointRecPo rec);
 
 
-    Long getIncreasedPoint(@Param(value = "pointTableName") String pointTableName,
-                           @Param(value = "uid") String uid,
-                           @Param(value = "start") Date start,
-                           @Param(value = "end") Date end);
+    BigDecimal getIncreasedPoint(@Param(value = "pointTableName") String pointTableName,
+                                 @Param(value = "uid") String uid,
+                                 @Param(value = "start") Date start,
+                                 @Param(value = "end") Date end);
 
     void batchUpdate(@Param(value = "pointTableName") String pointTableName,
                          @Param(value = "list") List<PointRecPo> list);
