+++ hyena-core/src/main/java/io/github/alphajiang/hyena/model/po/PointRecLogPo.java
@@ -24,6 +24,8 @@ import lombok.EqualsAndHashCode;
 import lombok.ToString;
 import lombok.experimental.Accessors;
 
+import java.math.BigDecimal;
+
 @Data
 @Accessors(chain = true)
 @EqualsAndHashCode(callSuper = true)
@@ -38,23 +40,23 @@ public class PointRecLogPo extends BasePo {
      * PointOpType
      */
     private Integer type;
-    private Long delta;
+    private BigDecimal delta;
     @ApiModelProperty("变动部分的实际成本")
-    private Long deltaCost;
-    private Long available;
-    private Long used;
-    private Long frozen;
-    private Long refund;
-    private Long cancelled;
-    private Long expire;
+    private BigDecimal deltaCost;
+    private BigDecimal available;
+    private BigDecimal used;
+    private BigDecimal frozen;
+    private BigDecimal refund;
+    private BigDecimal cancelled;
+    private BigDecimal expire;
     @ApiModelProperty("变动后,实际成本")
-    private Long cost;
+    private BigDecimal cost;
     @ApiModelProperty("变动后,冻结的实际成本")
-    private Long frozenCost;
+    private BigDecimal frozenCost;
     @ApiModelProperty("变动后,已消耗的实际成本")
-    private Long usedCost;
+    private BigDecimal usedCost;
     @ApiModelProperty("已退款的成本")
-    private Long refundCost;
+    private BigDecimal refundCost;
 
     private Integer sourceType;
     private Integer orderType;
